requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
onnxruntime==1.20.0
rembg==2.0.62
asyncpg==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
